Date - 14/11/20
I am not sure we need this as a seperate class. I have added this code to  the TestAldrian java.
I am not sure that search.java is correct use of the concept of classes by convention. But let me know what you all think?
It is a great piece of code and adds real functionality to the program