Most up to date as of 14/11/20
Alistair, I love this code. I think we should take what we need of it from here and add to TestAldrian class, We can rename this TestAldrian Customer order or something to that effect. I have directed a few options from the main menu method to your code.
We have set up an arrayList for  all the rentals and a separate parallel RentalsID arrayList which each order should be added to on instantiating the order object.
let me know if you have any questions